role-b access mw

+++ backend/routes/mcriap.js
@@ -66,6 +66,8 @@ router.get("/:id", async (req, res) => {
 });
 
 router.put("/:id", async (req, res) => {
+  if (req.role !== "manager")
+    return res.status(403).json({ message: "Forbidden" });
   try {
     const id = Number(req.params.id);
     const data = req.body;
@@ -101,6 +103,8 @@ router.put("/:id", async (req, res) => {
 });
 
 router.post("/", async (req, res) => {
+  if (req.role !== "manager")
+    return res.status(403).json({ message: "Forbidden" });
   try {
     const newChannel = await prisma.mcriapChannel.create({
       data: {
@@ -130,6 +134,8 @@ router.post("/", async (req, res) => {
 });
 
 router.delete("/:id", async (req, res) => {
+  if (req.role !== "manager")
+    return res.status(403).json({ message: "Forbidden" });
   try {
     const id = Number(req.params.id);
     await prisma.mcriapChannel.delete({
